+++ src/gameobjects/movable/car/PlayerCar.java
@@ -18,7 +18,7 @@ public class PlayerCar extends Car {
     }
 
     @Override
-    public void update(long duration) {
+    public void update(double duration) {
         location.setLocation(getX(), getY()+velocity);
     }
